requirements.txt
asgiref==3.8.1
Django==5.1.1
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-user-agents==0.4.0
gunicorn==23.0.0
packaging==24.1
sqlparse==0.5.1
typing_extensions==4.12.2
ua-parser==0.18.0
user-agents==2.2.0
